bs: Update project tree [3/3]
