+++ src/toollib/ap/ap.c
@@ -1,4 +1,4 @@
-#include <toollib/ap/ap.h>
+#include <toollib/ap.h>
 
 #include <string.h>
