+++ src/toollib/carea/carea.c
@@ -1,4 +1,4 @@
-#include <toollib/carea/carea.h>
+#include <toollib/carea.h>
 
 #include <malloc.h>
 #include <string.h>
