+++ src/toollib/csv/encoding.c
@@ -1,4 +1,4 @@
-#include <toollib/csv/csv.h>
+#include <toollib/csv.h>
 
 #include <stdlib.h>
