+++ src/toollib/csv/parseRow.c
@@ -1,4 +1,4 @@
-#include <toollib/csv/csv.h>
+#include <toollib/csv.h>
 
 #include <stdlib.h>
 #include <string.h>
