+++ test/toollib/ap++.cpp
@@ -2,17 +2,17 @@
 #include <sstream>
 #include <vector>
 
-#include <toollib/ap/ap.hpp>
+#include <toollib/ap.hpp>
 
 namespace tl = toollib;
 
 struct Case{
 	std::vector<const char*> args;
 	std::vector<std::string> lines;
-	tl::ap::ArgStatus endStatus = tl::ap::ArgStatus::OK;
+	tl::ap::ArgStatus endStatus = tl::ap::ArgStatus::EOI;
 };
 
-std::vector<Case> cases = {
+const std::vector<Case> cases = {
 	{
 		{"-"},
 		{
@@ -38,13 +38,15 @@ std::vector<Case> cases = {
 	{
 		{"a"},
 		{
-			"Arg[-1]: \"a\" \"(null)\""
+			"Arg[-1]: \"a\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"-a"},
 		{
+			"Arg[a]: \"(null)\" \"(null)\"",
 			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
@@ -59,49 +61,56 @@ std::vector<Case> cases = {
 	{
 		{"-b:v1"},
 		{
-			"Arg[ab]: \"v1\" \"(null)\""
+			"Arg[ab]: \"v1\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"-b=v1"},
 		{
-			"Arg[ab]: \"v1\" \"(null)\""
+			"Arg[ab]: \"v1\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"-bv1"},
 		{
-			"Arg[ab]: \"v1\" \"(null)\""
+			"Arg[ab]: \"v1\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"-c"},
 		{
-			"Arg[ac]: \"(null)\" \"(null)\""
+			"Arg[ac]: \"(null)\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"-c:v1"},
 		{
-			"Arg[ac]: \"v1\" \"(null)\""
+			"Arg[ac]: \"v1\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"-c=v1"},
 		{
-			"Arg[ac]: \"v1\" \"(null)\""
+			"Arg[ac]: \"v1\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"-cv1"},
 		{
-			"Arg[ac]: \"v1\" \"(null)\""
+			"Arg[ac]: \"v1\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -129,28 +138,32 @@ std::vector<Case> cases = {
 	{
 		{"-d:v1:v2"},
 		{
-			"Arg[ad]: \"v1\" \"v2\""
+			"Arg[ad]: \"v1\" \"v2\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"-d:v1=v2"},
 		{
-			"Arg[ad]: \"v1\" \"v2\""
+			"Arg[ad]: \"v1\" \"v2\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"-d=v1:v2"},
 		{
-			"Arg[ad]: \"v1\" \"v2\""
+			"Arg[ad]: \"v1\" \"v2\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"-d=v1=v2"},
 		{
-			"Arg[ad]: \"v1\" \"v2\""
+			"Arg[ad]: \"v1\" \"v2\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -171,14 +184,16 @@ std::vector<Case> cases = {
 	{
 		{"-dv1:v2"},
 		{
-			"Arg[ad]: \"v1\" \"v2\""
+			"Arg[ad]: \"v1\" \"v2\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"-dv1=v2"},
 		{
-			"Arg[ad]: \"v1\" \"v2\""
+			"Arg[ad]: \"v1\" \"v2\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -192,69 +207,79 @@ std::vector<Case> cases = {
 	{
 		{"-e:v1"},
 		{
-			"Arg[ee]: \"v1\" \"(null)\""
+			"Arg[ee]: \"v1\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"-e=v1"},
 		{
-			"Arg[ee]: \"v1\" \"(null)\""
+			"Arg[ee]: \"v1\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"-ev1"},
 		{
-			"Arg[ee]: \"v1\" \"(null)\""
+			"Arg[ee]: \"v1\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"-e:v1:v2"},
 		{
-			"Arg[ee]: \"v1\" \"v2\""
+			"Arg[ee]: \"v1\" \"v2\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"-e:v1=v2"},
 		{
-			"Arg[ee]: \"v1\" \"v2\""
+			"Arg[ee]: \"v1\" \"v2\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"-e=v1:v2"},
 		{
-			"Arg[ee]: \"v1\" \"v2\""
+			"Arg[ee]: \"v1\" \"v2\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"-e=v1=v2"},
 		{
-			"Arg[ee]: \"v1\" \"v2\""
+			"Arg[ee]: \"v1\" \"v2\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"-ev1:v2"},
 		{
-			"Arg[ee]: \"v1\" \"v2\""
+			"Arg[ee]: \"v1\" \"v2\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"-ev1=v2"},
 		{
-			"Arg[ee]: \"v1\" \"v2\""
+			"Arg[ee]: \"v1\" \"v2\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"--a"},
 		{
+			"Arg[a]: \"(null)\" \"(null)\"",
 			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
@@ -269,35 +294,40 @@ std::vector<Case> cases = {
 	{
 		{"--ab:v1"},
 		{
-			"Arg[ab]: \"v1\" \"(null)\""
+			"Arg[ab]: \"v1\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"--ab=v1"},
 		{
-			"Arg[ab]: \"v1\" \"(null)\""
+			"Arg[ab]: \"v1\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"--ac"},
 		{
-			"Arg[ac]: \"(null)\" \"(null)\""
+			"Arg[ac]: \"(null)\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"--ac:v1"},
 		{
-			"Arg[ac]: \"v1\" \"(null)\""
+			"Arg[ac]: \"v1\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"--ac=v1"},
 		{
-			"Arg[ac]: \"v1\" \"(null)\""
+			"Arg[ac]: \"v1\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -325,28 +355,32 @@ std::vector<Case> cases = {
 	{
 		{"--ad:v1:v2"},
 		{
-			"Arg[ad]: \"v1\" \"v2\""
+			"Arg[ad]: \"v1\" \"v2\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"--ad:v1=v2"},
 		{
-			"Arg[ad]: \"v1\" \"v2\""
+			"Arg[ad]: \"v1\" \"v2\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"--ad=v1:v2"},
 		{
-			"Arg[ad]: \"v1\" \"v2\""
+			"Arg[ad]: \"v1\" \"v2\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"--ad=v1=v2"},
 		{
-			"Arg[ad]: \"v1\" \"v2\""
+			"Arg[ad]: \"v1\" \"v2\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -360,42 +394,48 @@ std::vector<Case> cases = {
 	{
 		{"--ee:v1"},
 		{
-			"Arg[ee]: \"v1\" \"(null)\""
+			"Arg[ee]: \"v1\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"--ee=v1"},
 		{
-			"Arg[ee]: \"v1\" \"(null)\""
+			"Arg[ee]: \"v1\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"--ee:v1:v2"},
 		{
-			"Arg[ee]: \"v1\" \"v2\""
+			"Arg[ee]: \"v1\" \"v2\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"--ee:v1=v2"},
 		{
-			"Arg[ee]: \"v1\" \"v2\""
+			"Arg[ee]: \"v1\" \"v2\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"--ee=v1:v2"},
 		{
-			"Arg[ee]: \"v1\" \"v2\""
+			"Arg[ee]: \"v1\" \"v2\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"--ee=v1=v2"},
 		{
-			"Arg[ee]: \"v1\" \"v2\""
+			"Arg[ee]: \"v1\" \"v2\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -403,7 +443,8 @@ std::vector<Case> cases = {
 		{"a", "b"},
 		{
 			"Arg[-1]: \"a\" \"(null)\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -411,14 +452,16 @@ std::vector<Case> cases = {
 		{"-a", "b"},
 		{
 			"Arg[a]: \"(null)\" \"(null)\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"-b", "b"},
 		{
-			"Arg[ab]: \"b\" \"(null)\""
+			"Arg[ab]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -426,7 +469,8 @@ std::vector<Case> cases = {
 		{"-b:v1", "b"},
 		{
 			"Arg[ab]: \"v1\" \"(null)\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -434,7 +478,8 @@ std::vector<Case> cases = {
 		{"-b=v1", "b"},
 		{
 			"Arg[ab]: \"v1\" \"(null)\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -442,7 +487,8 @@ std::vector<Case> cases = {
 		{"-bv1", "b"},
 		{
 			"Arg[ab]: \"v1\" \"(null)\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -450,7 +496,8 @@ std::vector<Case> cases = {
 		{"-c", "b"},
 		{
 			"Arg[ac]: \"(null)\" \"(null)\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -458,7 +505,8 @@ std::vector<Case> cases = {
 		{"-c:v1", "b"},
 		{
 			"Arg[ac]: \"v1\" \"(null)\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -466,7 +514,8 @@ std::vector<Case> cases = {
 		{"-c=v1", "b"},
 		{
 			"Arg[ac]: \"v1\" \"(null)\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -474,7 +523,8 @@ std::vector<Case> cases = {
 		{"-cv1", "b"},
 		{
 			"Arg[ac]: \"v1\" \"(null)\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -488,14 +538,16 @@ std::vector<Case> cases = {
 	{
 		{"-d:v1", "b"},
 		{
-			"Arg[ad]: \"v1\" \"b\""
+			"Arg[ad]: \"v1\" \"b\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"-d=v1", "b"},
 		{
-			"Arg[ad]: \"v1\" \"b\""
+			"Arg[ad]: \"v1\" \"b\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -503,7 +555,8 @@ std::vector<Case> cases = {
 		{"-d:v1:v2", "b"},
 		{
 			"Arg[ad]: \"v1\" \"v2\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -511,7 +564,8 @@ std::vector<Case> cases = {
 		{"-d:v1=v2", "b"},
 		{
 			"Arg[ad]: \"v1\" \"v2\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -519,7 +573,8 @@ std::vector<Case> cases = {
 		{"-d=v1:v2", "b"},
 		{
 			"Arg[ad]: \"v1\" \"v2\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -527,21 +582,24 @@ std::vector<Case> cases = {
 		{"-d=v1=v2", "b"},
 		{
 			"Arg[ad]: \"v1\" \"v2\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"-dv1", "b"},
 		{
-			"Arg[ad]: \"v1\" \"b\""
+			"Arg[ad]: \"v1\" \"b\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"-dv1", "b"},
 		{
-			"Arg[ad]: \"v1\" \"b\""
+			"Arg[ad]: \"v1\" \"b\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -549,7 +607,8 @@ std::vector<Case> cases = {
 		{"-dv1:v2", "b"},
 		{
 			"Arg[ad]: \"v1\" \"v2\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -557,14 +616,16 @@ std::vector<Case> cases = {
 		{"-dv1=v2", "b"},
 		{
 			"Arg[ad]: \"v1\" \"v2\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"-e", "b"},
 		{
-			"Arg[ee]: \"b\" \"(null)\""
+			"Arg[ee]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -572,7 +633,8 @@ std::vector<Case> cases = {
 		{"-e:v1", "b"},
 		{
 			"Arg[ee]: \"v1\" \"(null)\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -580,7 +642,8 @@ std::vector<Case> cases = {
 		{"-e=v1", "b"},
 		{
 			"Arg[ee]: \"v1\" \"(null)\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -588,7 +651,8 @@ std::vector<Case> cases = {
 		{"-ev1", "b"},
 		{
 			"Arg[ee]: \"v1\" \"(null)\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -596,7 +660,8 @@ std::vector<Case> cases = {
 		{"-e:v1:v2", "b"},
 		{
 			"Arg[ee]: \"v1\" \"v2\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -604,7 +669,8 @@ std::vector<Case> cases = {
 		{"-e:v1=v2", "b"},
 		{
 			"Arg[ee]: \"v1\" \"v2\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -612,7 +678,8 @@ std::vector<Case> cases = {
 		{"-e=v1:v2", "b"},
 		{
 			"Arg[ee]: \"v1\" \"v2\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -620,7 +687,8 @@ std::vector<Case> cases = {
 		{"-e=v1=v2", "b"},
 		{
 			"Arg[ee]: \"v1\" \"v2\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -628,7 +696,8 @@ std::vector<Case> cases = {
 		{"-ev1:v2", "b"},
 		{
 			"Arg[ee]: \"v1\" \"v2\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -636,7 +705,8 @@ std::vector<Case> cases = {
 		{"-ev1=v2", "b"},
 		{
 			"Arg[ee]: \"v1\" \"v2\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -644,14 +714,16 @@ std::vector<Case> cases = {
 		{"--a", "b"},
 		{
 			"Arg[a]: \"(null)\" \"(null)\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"--ab", "b"},
 		{
-			"Arg[ab]: \"b\" \"(null)\""
+			"Arg[ab]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -659,7 +731,8 @@ std::vector<Case> cases = {
 		{"--ab:v1", "b"},
 		{
 			"Arg[ab]: \"v1\" \"(null)\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -667,7 +740,8 @@ std::vector<Case> cases = {
 		{"--ab=v1", "b"},
 		{
 			"Arg[ab]: \"v1\" \"(null)\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -675,7 +749,8 @@ std::vector<Case> cases = {
 		{"--ac", "b"},
 		{
 			"Arg[ac]: \"(null)\" \"(null)\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -683,7 +758,8 @@ std::vector<Case> cases = {
 		{"--ac:v1", "b"},
 		{
 			"Arg[ac]: \"v1\" \"(null)\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -691,7 +767,8 @@ std::vector<Case> cases = {
 		{"--ac=v1", "b"},
 		{
 			"Arg[ac]: \"v1\" \"(null)\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -705,14 +782,16 @@ std::vector<Case> cases = {
 	{
 		{"--ad:v1", "b"},
 		{
-			"Arg[ad]: \"v1\" \"b\""
+			"Arg[ad]: \"v1\" \"b\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"--ad=v1", "b"},
 		{
-			"Arg[ad]: \"v1\" \"b\""
+			"Arg[ad]: \"v1\" \"b\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -720,7 +799,8 @@ std::vector<Case> cases = {
 		{"--ad:v1:v2", "b"},
 		{
 			"Arg[ad]: \"v1\" \"v2\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -728,7 +808,8 @@ std::vector<Case> cases = {
 		{"--ad:v1=v2", "b"},
 		{
 			"Arg[ad]: \"v1\" \"v2\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -736,7 +817,8 @@ std::vector<Case> cases = {
 		{"--ad=v1:v2", "b"},
 		{
 			"Arg[ad]: \"v1\" \"v2\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -744,14 +826,16 @@ std::vector<Case> cases = {
 		{"--ad=v1=v2", "b"},
 		{
 			"Arg[ad]: \"v1\" \"v2\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"--ee", "b"},
 		{
-			"Arg[ee]: \"b\" \"(null)\""
+			"Arg[ee]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -759,7 +843,8 @@ std::vector<Case> cases = {
 		{"--ee:v1", "b"},
 		{
 			"Arg[ee]: \"v1\" \"(null)\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -767,7 +852,8 @@ std::vector<Case> cases = {
 		{"--ee=v1", "b"},
 		{
 			"Arg[ee]: \"v1\" \"(null)\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -775,7 +861,8 @@ std::vector<Case> cases = {
 		{"--ee:v1:v2", "b"},
 		{
 			"Arg[ee]: \"v1\" \"v2\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -783,7 +870,8 @@ std::vector<Case> cases = {
 		{"--ee:v1=v2", "b"},
 		{
 			"Arg[ee]: \"v1\" \"v2\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -791,7 +879,8 @@ std::vector<Case> cases = {
 		{"--ee=v1:v2", "b"},
 		{
 			"Arg[ee]: \"v1\" \"v2\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -799,7 +888,8 @@ std::vector<Case> cases = {
 		{"--ee=v1=v2", "b"},
 		{
 			"Arg[ee]: \"v1\" \"v2\"",
-			"Arg[-1]: \"b\" \"(null)\""
+			"Arg[-1]: \"b\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -808,7 +898,8 @@ std::vector<Case> cases = {
 		{"a", "v1:v2"},
 		{
 			"Arg[-1]: \"a\" \"(null)\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -816,14 +907,16 @@ std::vector<Case> cases = {
 		{"-a", "v1:v2"},
 		{
 			"Arg[a]: \"(null)\" \"(null)\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"-b", "v1:v2"},
 		{
-			"Arg[ab]: \"v1:v2\" \"(null)\""
+			"Arg[ab]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -831,7 +924,8 @@ std::vector<Case> cases = {
 		{"-b:v1", "v1:v2"},
 		{
 			"Arg[ab]: \"v1\" \"(null)\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -839,7 +933,8 @@ std::vector<Case> cases = {
 		{"-b=v1", "v1:v2"},
 		{
 			"Arg[ab]: \"v1\" \"(null)\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -847,7 +942,8 @@ std::vector<Case> cases = {
 		{"-bv1", "v1:v2"},
 		{
 			"Arg[ab]: \"v1\" \"(null)\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -855,7 +951,8 @@ std::vector<Case> cases = {
 		{"-c", "v1:v2"},
 		{
 			"Arg[ac]: \"(null)\" \"(null)\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -863,7 +960,8 @@ std::vector<Case> cases = {
 		{"-c:v1", "v1:v2"},
 		{
 			"Arg[ac]: \"v1\" \"(null)\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -871,7 +969,8 @@ std::vector<Case> cases = {
 		{"-c=v1", "v1:v2"},
 		{
 			"Arg[ac]: \"v1\" \"(null)\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -879,28 +978,32 @@ std::vector<Case> cases = {
 		{"-cv1", "v1:v2"},
 		{
 			"Arg[ac]: \"v1\" \"(null)\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"-d", "v1:v2"},
 		{
-			"Arg[ad]: \"v1\" \"v2\""
+			"Arg[ad]: \"v1\" \"v2\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"-d:v1", "v1:v2"},
 		{
-			"Arg[ad]: \"v1\" \"v1:v2\""
+			"Arg[ad]: \"v1\" \"v1:v2\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"-d=v1", "v1:v2"},
 		{
-			"Arg[ad]: \"v1\" \"v1:v2\""
+			"Arg[ad]: \"v1\" \"v1:v2\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -908,7 +1011,8 @@ std::vector<Case> cases = {
 		{"-d:v1:v2", "v1:v2"},
 		{
 			"Arg[ad]: \"v1\" \"v2\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -916,7 +1020,8 @@ std::vector<Case> cases = {
 		{"-d:v1=v2", "v1:v2"},
 		{
 			"Arg[ad]: \"v1\" \"v2\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -924,7 +1029,8 @@ std::vector<Case> cases = {
 		{"-d=v1:v2", "v1:v2"},
 		{
 			"Arg[ad]: \"v1\" \"v2\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -932,21 +1038,24 @@ std::vector<Case> cases = {
 		{"-d=v1=v2", "v1:v2"},
 		{
 			"Arg[ad]: \"v1\" \"v2\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"-dv1", "v1:v2"},
 		{
-			"Arg[ad]: \"v1\" \"v1:v2\""
+			"Arg[ad]: \"v1\" \"v1:v2\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"-dv1", "v1:v2"},
 		{
-			"Arg[ad]: \"v1\" \"v1:v2\""
+			"Arg[ad]: \"v1\" \"v1:v2\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -954,7 +1063,8 @@ std::vector<Case> cases = {
 		{"-dv1:v2", "v1:v2"},
 		{
 			"Arg[ad]: \"v1\" \"v2\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -962,14 +1072,16 @@ std::vector<Case> cases = {
 		{"-dv1=v2", "v1:v2"},
 		{
 			"Arg[ad]: \"v1\" \"v2\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"-e", "v1:v2"},
 		{
-			"Arg[ee]: \"v1\" \"v2\""
+			"Arg[ee]: \"v1\" \"v2\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -977,7 +1089,8 @@ std::vector<Case> cases = {
 		{"-e:v1", "v1:v2"},
 		{
 			"Arg[ee]: \"v1\" \"(null)\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -985,7 +1098,8 @@ std::vector<Case> cases = {
 		{"-e=v1", "v1:v2"},
 		{
 			"Arg[ee]: \"v1\" \"(null)\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -993,7 +1107,8 @@ std::vector<Case> cases = {
 		{"-ev1", "v1:v2"},
 		{
 			"Arg[ee]: \"v1\" \"(null)\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1001,7 +1116,8 @@ std::vector<Case> cases = {
 		{"-e:v1:v2", "v1:v2"},
 		{
 			"Arg[ee]: \"v1\" \"v2\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1009,7 +1125,8 @@ std::vector<Case> cases = {
 		{"-e:v1=v2", "v1:v2"},
 		{
 			"Arg[ee]: \"v1\" \"v2\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1017,7 +1134,8 @@ std::vector<Case> cases = {
 		{"-e=v1:v2", "v1:v2"},
 		{
 			"Arg[ee]: \"v1\" \"v2\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1025,7 +1143,8 @@ std::vector<Case> cases = {
 		{"-e=v1=v2", "v1:v2"},
 		{
 			"Arg[ee]: \"v1\" \"v2\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1033,7 +1152,8 @@ std::vector<Case> cases = {
 		{"-ev1:v2", "v1:v2"},
 		{
 			"Arg[ee]: \"v1\" \"v2\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1041,7 +1161,8 @@ std::vector<Case> cases = {
 		{"-ev1=v2", "v1:v2"},
 		{
 			"Arg[ee]: \"v1\" \"v2\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1049,14 +1170,16 @@ std::vector<Case> cases = {
 		{"--a", "v1:v2"},
 		{
 			"Arg[a]: \"(null)\" \"(null)\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"--ab", "v1:v2"},
 		{
-			"Arg[ab]: \"v1:v2\" \"(null)\""
+			"Arg[ab]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1064,7 +1187,8 @@ std::vector<Case> cases = {
 		{"--ab:v1", "v1:v2"},
 		{
 			"Arg[ab]: \"v1\" \"(null)\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1072,7 +1196,8 @@ std::vector<Case> cases = {
 		{"--ab=v1", "v1:v2"},
 		{
 			"Arg[ab]: \"v1\" \"(null)\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1080,7 +1205,8 @@ std::vector<Case> cases = {
 		{"--ac", "v1:v2"},
 		{
 			"Arg[ac]: \"(null)\" \"(null)\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1088,7 +1214,8 @@ std::vector<Case> cases = {
 		{"--ac:v1", "v1:v2"},
 		{
 			"Arg[ac]: \"v1\" \"(null)\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1096,28 +1223,32 @@ std::vector<Case> cases = {
 		{"--ac=v1", "v1:v2"},
 		{
 			"Arg[ac]: \"v1\" \"(null)\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"--ad", "v1:v2"},
 		{
-			"Arg[ad]: \"v1\" \"v2\""
+			"Arg[ad]: \"v1\" \"v2\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"--ad:v1", "v1:v2"},
 		{
-			"Arg[ad]: \"v1\" \"v1:v2\""
+			"Arg[ad]: \"v1\" \"v1:v2\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"--ad=v1", "v1:v2"},
 		{
-			"Arg[ad]: \"v1\" \"v1:v2\""
+			"Arg[ad]: \"v1\" \"v1:v2\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1125,7 +1256,8 @@ std::vector<Case> cases = {
 		{"--ad:v1:v2", "v1:v2"},
 		{
 			"Arg[ad]: \"v1\" \"v2\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1133,7 +1265,8 @@ std::vector<Case> cases = {
 		{"--ad:v1=v2", "v1:v2"},
 		{
 			"Arg[ad]: \"v1\" \"v2\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1141,7 +1274,8 @@ std::vector<Case> cases = {
 		{"--ad=v1:v2", "v1:v2"},
 		{
 			"Arg[ad]: \"v1\" \"v2\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1149,14 +1283,16 @@ std::vector<Case> cases = {
 		{"--ad=v1=v2", "v1:v2"},
 		{
 			"Arg[ad]: \"v1\" \"v2\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"--ee", "v1:v2"},
 		{
-			"Arg[ee]: \"v1\" \"v2\""
+			"Arg[ee]: \"v1\" \"v2\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1164,7 +1300,8 @@ std::vector<Case> cases = {
 		{"--ee:v1", "v1:v2"},
 		{
 			"Arg[ee]: \"v1\" \"(null)\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1172,7 +1309,8 @@ std::vector<Case> cases = {
 		{"--ee=v1", "v1:v2"},
 		{
 			"Arg[ee]: \"v1\" \"(null)\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1180,7 +1318,8 @@ std::vector<Case> cases = {
 		{"--ee:v1:v2", "v1:v2"},
 		{
 			"Arg[ee]: \"v1\" \"v2\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1188,7 +1327,8 @@ std::vector<Case> cases = {
 		{"--ee:v1=v2", "v1:v2"},
 		{
 			"Arg[ee]: \"v1\" \"v2\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1196,7 +1336,8 @@ std::vector<Case> cases = {
 		{"--ee=v1:v2", "v1:v2"},
 		{
 			"Arg[ee]: \"v1\" \"v2\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1204,7 +1345,8 @@ std::vector<Case> cases = {
 		{"--ee=v1=v2", "v1:v2"},
 		{
 			"Arg[ee]: \"v1\" \"v2\"",
-			"Arg[-1]: \"v1:v2\" \"(null)\""
+			"Arg[-1]: \"v1:v2\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1213,7 +1355,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[-1]: \"a\" \"(null)\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1222,7 +1365,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[a]: \"(null)\" \"(null)\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1230,7 +1374,8 @@ std::vector<Case> cases = {
 		{"-b", "b", "c"},
 		{
 			"Arg[ab]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1239,7 +1384,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ab]: \"v1\" \"(null)\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1248,7 +1394,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ab]: \"v1\" \"(null)\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1257,7 +1404,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ab]: \"v1\" \"(null)\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1266,7 +1414,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ac]: \"(null)\" \"(null)\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1275,7 +1424,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ac]: \"v1\" \"(null)\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1284,7 +1434,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ac]: \"v1\" \"(null)\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1293,14 +1444,16 @@ std::vector<Case> cases = {
 		{
 			"Arg[ac]: \"v1\" \"(null)\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"-d", "b", "c"},
 		{
-			"Arg[ad]: \"b\" \"c\""
+			"Arg[ad]: \"b\" \"c\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1308,7 +1461,8 @@ std::vector<Case> cases = {
 		{"-d:v1", "b", "c"},
 		{
 			"Arg[ad]: \"v1\" \"b\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1316,7 +1470,8 @@ std::vector<Case> cases = {
 		{"-d=v1", "b", "c"},
 		{
 			"Arg[ad]: \"v1\" \"b\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1325,7 +1480,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ad]: \"v1\" \"v2\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1334,7 +1490,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ad]: \"v1\" \"v2\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1343,7 +1500,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ad]: \"v1\" \"v2\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1352,7 +1510,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ad]: \"v1\" \"v2\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1360,7 +1519,8 @@ std::vector<Case> cases = {
 		{"-dv1", "b", "c"},
 		{
 			"Arg[ad]: \"v1\" \"b\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1368,7 +1528,8 @@ std::vector<Case> cases = {
 		{"-dv1", "b", "c"},
 		{
 			"Arg[ad]: \"v1\" \"b\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1377,7 +1538,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ad]: \"v1\" \"v2\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1386,7 +1548,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ad]: \"v1\" \"v2\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1394,7 +1557,8 @@ std::vector<Case> cases = {
 		{"-e", "b", "c"},
 		{
 			"Arg[ee]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1403,7 +1567,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ee]: \"v1\" \"(null)\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1412,7 +1577,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ee]: \"v1\" \"(null)\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1421,7 +1587,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ee]: \"v1\" \"(null)\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1430,7 +1597,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ee]: \"v1\" \"v2\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1439,7 +1607,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ee]: \"v1\" \"v2\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1448,7 +1617,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ee]: \"v1\" \"v2\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1457,7 +1627,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ee]: \"v1\" \"v2\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1466,7 +1637,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ee]: \"v1\" \"v2\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1475,7 +1647,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ee]: \"v1\" \"v2\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1484,7 +1657,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[a]: \"(null)\" \"(null)\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1492,7 +1666,8 @@ std::vector<Case> cases = {
 		{"--ab", "b", "c"},
 		{
 			"Arg[ab]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1501,7 +1676,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ab]: \"v1\" \"(null)\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1510,7 +1686,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ab]: \"v1\" \"(null)\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1519,7 +1696,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ac]: \"(null)\" \"(null)\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1528,7 +1706,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ac]: \"v1\" \"(null)\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1537,14 +1716,16 @@ std::vector<Case> cases = {
 		{
 			"Arg[ac]: \"v1\" \"(null)\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
 	{
 		{"--ad", "b", "c"},
 		{
-			"Arg[ad]: \"b\" \"c\""
+			"Arg[ad]: \"b\" \"c\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1552,7 +1733,8 @@ std::vector<Case> cases = {
 		{"--ad:v1", "b", "c"},
 		{
 			"Arg[ad]: \"v1\" \"b\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1560,7 +1742,8 @@ std::vector<Case> cases = {
 		{"--ad=v1", "b", "c"},
 		{
 			"Arg[ad]: \"v1\" \"b\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1569,7 +1752,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ad]: \"v1\" \"v2\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1578,7 +1762,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ad]: \"v1\" \"v2\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1587,7 +1772,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ad]: \"v1\" \"v2\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1596,7 +1782,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ad]: \"v1\" \"v2\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1604,7 +1791,8 @@ std::vector<Case> cases = {
 		{"--ee", "b", "c"},
 		{
 			"Arg[ee]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1613,7 +1801,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ee]: \"v1\" \"(null)\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1622,7 +1811,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ee]: \"v1\" \"(null)\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1631,7 +1821,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ee]: \"v1\" \"v2\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1640,7 +1831,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ee]: \"v1\" \"v2\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1649,7 +1841,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ee]: \"v1\" \"v2\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1658,7 +1851,8 @@ std::vector<Case> cases = {
 		{
 			"Arg[ee]: \"v1\" \"v2\"",
 			"Arg[-1]: \"b\" \"(null)\"",
-			"Arg[-1]: \"c\" \"(null)\""
+			"Arg[-1]: \"c\" \"(null)\"",
+			"Arg[a]: \"(null)\" \"(null)\""
 		},
 		tl::ap::ArgStatus::EOI
 	},
@@ -1674,21 +1868,19 @@ int main(){
 		{ 'e', "ee", tl::ap::ArgType::VALUE2_OPTIONAL }
 	};
 
+	tl::ap::Arg arg;
 	size_t passed = 0;
+
 	for(size_t casei = 0; casei < cases.size(); casei++){
-		size_t lpassed = 0;
-		
 		int argc = cases[casei].args.size();
 		const char* const* argv = cases[casei].args.data();
-		auto args = tl::ap::getAll(5, descs + 1, &argc, &argv);
+		size_t lpassed = 0;
 
-		tl::ap::ArgStatus es = tl::ap::ArgStatus::EOI;
-		for(size_t i = 0; i < args.size(); i++){
-			auto& arg = args[i];
-			auto& line = cases[casei].lines[i];
-
+		for(auto& line: cases[casei].lines){
 			std::stringstream ss;
-
+		
+			arg = tl::ap::next(5, descs + 1, &argc, &argv);
+		
 			ss << "Arg[" << descs[arg.m_Id + 1].m_Long << "]: \"";
 			if(arg.m_Value && arg.m_ValueLen) 
 				ss << std::string(arg.m_Value).substr(0, arg.m_ValueLen); 
@@ -1706,18 +1898,17 @@ int main(){
 			else
 				std::cerr << "FAILED(case " << casei << "): Lines dismatch" << std::endl << "  G: " << ss.str() << std::endl << "  E: " << line << std::endl;
 		
-			if(arg.m_Status != tl::ap::ArgStatus::OK){
-				es = arg.m_Status;
+			if(arg.m_Status != tl::ap::ArgStatus::OK)
 				break;
-			}
 		}
-
-		if(lpassed == cases[casei].lines.size() && es == cases[casei].endStatus)
+		
+		if(lpassed == cases[casei].lines.size() && arg.m_Status == cases[casei].endStatus)
 			passed++;
 		else
-			std::cerr << "FAILED(case " << casei << "): End status dismatch" << std::endl << "  G: " << lpassed << ":" << (int)es << std::endl << "  E: " << cases[casei].lines.size() << ":" << (int)cases[casei].endStatus << std::endl;
+			std::cerr << "FAILED(case " << casei << "): End status dismatch" << std::endl << "  G: " << lpassed << ":" << (int)arg.m_Status << std::endl << "  E: " << cases[casei].lines.size() << ":" << (int)cases[casei].endStatus << std::endl;
 	}
-	
+
+		
 	if(passed == cases.size())
 		std::cout << "OK" << std::endl; 
 	else
