+++ test/toollib/carea.c
@@ -1,4 +1,4 @@
-#include <toollib/carea/carea.h>
+#include <toollib/carea.h>
 
 #include <stdio.h>
